bisect_make will ignore updates to the template
When git checks out files, it sets their modification time to the
present. So if git bisect checks out a commit with a different version
of the template, then the template will be newer than the local instance
of it, and causing the build to fail. Since we expect to see recently
modified files in the process of a make, the only sensible thing for the
"bisect_make" rule to do is ignore them, without even outputting a
warning.