Fix bug in .depends file generation
When "gcc -MM" generates rules that are sufficiently long, they are
wrapped as in this example:

target: prerequisite0 prerequisite1 prerequisite2 prerequisite3 \
     prerequisite4

Prior to this commit I used 'sed' to prepend directories like $(BIN_DIR)
to each line, not realizing that some rules would wrap. This resulted in
rules having dependencies on entire build directories, which made them
run unnecessarily.

This commit tweaks the makefile so that sed prepends text only to lines
that begin with a non-whitespace character.